requirements.txt
flask==3.0.3
redis==5.0.4
openai==1.30.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-mock==3.14.0
groq==0.9.0